somewhat working version (dirty)

+++ COPL2/main.cpp
@@ -38,15 +38,19 @@ int main(int argc, char* argv[]) {
         }
 
         parser.maakBoom();
-        parser.printBoom(parser.treeRoot);
-        // parser.preOrder(parser.treeRoot);
-        // parser.treeRoot = parser.postOrder(parser.treeRoot);
+        parser.printBoom(parser.treeRoot); 
+        parser.treeRoot = parser.postOrder(parser.treeRoot); 
+        if (parser.treeRoot != nullptr)
+        {
+            std::cout << std::endl << "result main" << std::endl;
+            parser.printBoom(parser.treeRoot);
+            std::cout << std::endl << "result main" << std::endl;
+            parser.deleteSubtree(parser.treeRoot);
+            parser.leegVector();
+            // exit(0);
+        }
+            
 
-        std::cout << std::endl << "result main" << std::endl;
-        parser.printBoom(parser.treeRoot);
-        std::cout << std::endl << "result main" << std::endl;
-        parser.deleteSubtree(parser.treeRoot);
-        parser.freeVector();
     }
 
     return 0;
